refactor: split plugins.go into domain specific go files and move it to internal
This is for consistency with other commands and their layouts
