+++ cmd/d8/root.go
@@ -37,14 +37,14 @@ import (
 	dkplog "github.com/deckhouse/deckhouse/pkg/log"
 
 	"github.com/deckhouse/deckhouse-cli/cmd/commands"
-	"github.com/deckhouse/deckhouse-cli/cmd/plugins"
-	"github.com/deckhouse/deckhouse-cli/cmd/plugins/flags"
 	backup "github.com/deckhouse/deckhouse-cli/internal/backup/cmd"
 	data "github.com/deckhouse/deckhouse-cli/internal/data/cmd"
 	iam "github.com/deckhouse/deckhouse-cli/internal/iam/cmd"
 	iamuser "github.com/deckhouse/deckhouse-cli/internal/iam/user/cmd"
 	mirror "github.com/deckhouse/deckhouse-cli/internal/mirror/cmd"
-	"github.com/deckhouse/deckhouse-cli/internal/network"
+	network "github.com/deckhouse/deckhouse-cli/internal/network"
+	plugins "github.com/deckhouse/deckhouse-cli/internal/plugins/cmd"
+	"github.com/deckhouse/deckhouse-cli/internal/plugins/cmd/flags"
 	status "github.com/deckhouse/deckhouse-cli/internal/status/cmd"
 	system "github.com/deckhouse/deckhouse-cli/internal/system/cmd"
 	"github.com/deckhouse/deckhouse-cli/internal/tools"
